requirements.txt
aioredis==1.3.1
asgiref==3.3.0
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.3
django-filter==2.4.0
django-heroku==0.3.1
django-mathfilters==1.0.0
django-multiselectfield==0.1.12
django-polymorphic==3.0.0
django-tables2==2.4.0
django-toolbelt==0.0.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
gunicorn==20.0.4
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
Markdown==3.3.4
msgpack==1.0.2
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pytz==2020.4
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
static3==0.7.0
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
whitenoise==5.2.0
zope.interface==5.4.0
